simage/ui/requirements.txt

PySide6
Pillow
pytest
